refactor: extract convert functions to folder
