+++ slime/backends/megatron_utils/update_weight_utils.py
@@ -6,6 +6,7 @@ from megatron.core import mpu
 from megatron.core.transformer.transformer_layer import get_transformer_layer_offset
 from slime.utils.types import ParamInfo
 from .initialize import get_gloo_group
+from .megatron_to_hf import convert_to_hf  # noqa: F401
 
 
 def all_gather_param(name, param):
@@ -197,545 +198,3 @@ def get_param_info_buckets(args, model) -> list[list[ParamInfo]]:
         param_info_buckets[-1].append(info)
         buffer_size += param_size
     return param_info_buckets
-
-
-def quantize_param(name, weight, weight_block_size):
-    assert name.endswith(".weight"), f"Expected weight parameter, got {name}"
-    FP8_MIN = torch.finfo(torch.float8_e4m3fn).min
-    FP8_MAX = torch.finfo(torch.float8_e4m3fn).max
-    if weight_block_size is not None:
-        # per block quant
-        block_n, block_k = weight_block_size[0], weight_block_size[1]
-
-        n_tiles = weight.shape[0] // block_n
-        k_tiles = weight.shape[1] // block_k
-
-        qweight = weight.reshape(n_tiles, block_n, k_tiles, block_k)
-        block_max = torch.max(torch.abs(qweight), dim=1, keepdim=True)[0]
-        block_max = torch.max(block_max, dim=3, keepdim=True)[0]
-
-        scale = block_max.to(torch.float32) / FP8_MIN
-        qweight = (qweight / scale).clamp(min=FP8_MIN, max=FP8_MAX).reshape(weight.shape).to(torch.float8_e4m3fn)
-        scale = scale.squeeze()
-        scale_name = name.replace(".weight", ".weight_scale_inv")
-    else:
-        # per tensor quant
-        scale = weight.abs().max().clamp(min=1e-12).to(torch.float32) / FP8_MAX
-        qweight = (weight / scale).clamp(min=FP8_MIN, max=FP8_MAX).to(torch.float8_e4m3fn)
-        scale = scale.view(1)
-        scale_name = name.replace(".weight", ".weight_scale")
-    return [(name, qweight), (scale_name, scale)]
-
-
-def quantize_params(args, megatron_name, converted_named_params, quantization_config):
-    if quantization_config is None:
-        return converted_named_params
-    assert quantization_config["quant_method"] == "fp8"
-    assert quantization_config["fmt"] == "e4m3"
-    assert quantization_config["activation_scheme"] == "dynamic"
-    weight_block_size = quantization_config.get("weight_block_size", None)
-
-    decoder_layers_pattern = r"module\.module\.decoder\.layers\.(\d+)\.(.+)"
-    match = re.match(decoder_layers_pattern, megatron_name)
-
-    if not match:
-        return converted_named_params
-
-    layer_idx, rest = match.groups()
-    # experts
-    expert_pattern = r"mlp.experts\.(.+)\.weight(\d+)"
-    match = re.match(expert_pattern, rest)
-    if match:
-        rest, expert_idx = match.groups()
-        if rest in [
-            "linear_fc1",
-            "linear_fc2",
-        ]:
-            quantize_named_params = []
-            for converted_name, param in converted_named_params:
-                # skip bf16 weight_scale and input_scale
-                # TODO: find a clearer way.
-                if converted_name.endswith("_scale"):
-                    continue
-                quantize_named_params.extend(quantize_param(converted_name, param, weight_block_size))
-
-            return quantize_named_params
-
-    # shared expert
-    shared_expert_pattern = r"mlp.shared_experts\.(.+)"
-    match = re.match(shared_expert_pattern, rest)
-    if match:
-        rest = match.groups()[0]
-        if rest in [
-            "linear_fc1.weight",
-            "linear_fc2.weight",
-        ]:
-            quantize_named_params = []
-            for converted_name, param in converted_named_params:
-                quantize_named_params.extend(quantize_param(converted_name, param, weight_block_size))
-
-            return quantize_named_params
-
-    if rest in [
-        "self_attention.linear_proj.weight",
-        "self_attention.linear_qkv.weight",
-        "mlp.linear_fc1.weight",
-        "mlp.linear_fc2.weight",
-    ]:
-        quantize_named_params = []
-        for converted_name, param in converted_named_params:
-            quantize_named_params.extend(quantize_param(converted_name, param, weight_block_size))
-
-        return quantize_named_params
-
-    # for other parameters, we just return the original converted_named_params
-    return converted_named_params
-
-
-def convert_to_hf(args, model_name, name, param, quantization_config=None):
-    if "glm4" in model_name:
-        converted_named_tensors = convert_glm4_to_hf(args, name, param)
-    elif "qwen3moe" in model_name:
-        converted_named_tensors = convert_qwen3moe_to_hf(args, name, param)
-    elif "qwen2" in model_name or "qwen3" in model_name:
-        converted_named_tensors = convert_qwen2_to_hf(args, name, param)
-    elif "deepseekv3" in model_name:
-        converted_named_tensors = convert_deepseekv3_to_hf(args, name, param)
-    elif "llama" in model_name:
-        converted_named_tensors = convert_llama_to_hf(args, name, param)
-    else:
-        raise ValueError(f"Unsupported model: {model_name}")
-
-    if not quantization_config:
-        return converted_named_tensors
-
-    return quantize_params(args, name, converted_named_tensors, quantization_config)
-
-
-def convert_glm4_to_hf(args, name, param):
-    if name == "module.module.embedding.word_embeddings.weight":
-        return [("model.embed_tokens.weight", param)]
-    if name == "module.module.output_layer.weight":
-        return [("lm_head.weight", param)]
-    if name == "module.module.decoder.final_layernorm.weight":
-        return [("model.norm.weight", param)]
-
-    try:
-        head_dim = args.kv_channels if args.kv_channels is not None else args.hidden_size // args.num_attention_heads
-    except:
-        head_dim = args.hidden_size // args.num_attention_heads
-    value_num_per_group = args.num_attention_heads // args.num_query_groups
-
-    decoder_layers_pattern = r"module\.module\.decoder\.layers\.(\d+)\.(.+)"
-    match = re.match(decoder_layers_pattern, name)
-    if match:
-        layer_idx, rest = match.groups()
-        if rest == "self_attention.linear_proj.weight":
-            return [(f"model.layers.{layer_idx}.self_attn.o_proj.weight", param)]
-        elif rest == "self_attention.linear_qkv.weight":
-
-            param = param.view(args.num_query_groups, -1, head_dim, args.hidden_size)
-            q_param, k_param, v_param = torch.split(param, split_size_or_sections=[value_num_per_group, 1, 1], dim=1)
-            q_param = q_param.reshape(-1, args.hidden_size)
-            k_param = k_param.reshape(-1, args.hidden_size)
-            v_param = v_param.reshape(-1, args.hidden_size)
-            return [
-                (f"model.layers.{layer_idx}.self_attn.q_proj.weight", q_param),
-                (f"model.layers.{layer_idx}.self_attn.k_proj.weight", k_param),
-                (f"model.layers.{layer_idx}.self_attn.v_proj.weight", v_param),
-            ]
-        elif rest == "self_attention.linear_qkv.bias":
-            param = param.view(args.num_query_groups, -1)
-            q_bias, k_bias, v_bias = torch.split(
-                param,
-                split_size_or_sections=[value_num_per_group * head_dim, head_dim, head_dim],
-                dim=1,
-            )
-            q_bias = q_bias.contiguous().flatten()
-            k_bias = k_bias.contiguous().flatten()
-            v_bias = v_bias.contiguous().flatten()
-            return [
-                (f"model.layers.{layer_idx}.self_attn.q_proj.bias", q_bias),
-                (f"model.layers.{layer_idx}.self_attn.k_proj.bias", k_bias),
-                (f"model.layers.{layer_idx}.self_attn.v_proj.bias", v_bias),
-            ]
-        elif rest == "mlp.linear_fc1.weight":
-            return [
-                (f"model.layers.{layer_idx}.mlp.gate_up_proj.weight", param),
-            ]
-        elif rest == "mlp.linear_fc2.weight":
-            return [(f"model.layers.{layer_idx}.mlp.down_proj.weight", param)]
-        elif rest == "self_attention.linear_qkv.layer_norm_weight":
-            return [(f"model.layers.{layer_idx}.input_layernorm.weight", param)]
-        elif rest == "mlp.linear_fc1.layer_norm_weight":
-            return [(f"model.layers.{layer_idx}.post_attention_layernorm.weight", param)]
-
-        # qk norm
-        elif rest == "self_attention.q_layernorm.weight":
-            return [(f"model.layers.{layer_idx}.self_attn.q_norm.weight", param)]
-        elif rest == "self_attention.k_layernorm.weight":
-            return [(f"model.layers.{layer_idx}.self_attn.k_norm.weight", param)]
-
-        # sandwitch norm
-        elif rest == "post_self_attn_layernorm.weight":
-            return [(f"model.layers.{layer_idx}.post_self_attn_layernorm.weight", param)]
-        elif rest == "post_mlp_layernorm.weight":
-            return [(f"model.layers.{layer_idx}.post_mlp_layernorm.weight", param)]
-
-    raise ValueError(f"Unknown parameter name: {name}")
-
-
-def convert_qwen2_to_hf(args, name, param):
-    if name == "module.module.embedding.word_embeddings.weight":
-        return [("model.embed_tokens.weight", param)]
-    if name == "module.module.output_layer.weight":
-        return [("lm_head.weight", param)]
-    if name == "module.module.decoder.final_layernorm.weight":
-        return [("model.norm.weight", param)]
-
-    try:
-        head_dim = args.kv_channels if args.kv_channels is not None else args.hidden_size // args.num_attention_heads
-    except:
-        head_dim = args.hidden_size // args.num_attention_heads
-    value_num_per_group = args.num_attention_heads // args.num_query_groups
-
-    decoder_layers_pattern = r"module\.module\.decoder\.layers\.(\d+)\.(.+)"
-    match = re.match(decoder_layers_pattern, name)
-    if match:
-        layer_idx, rest = match.groups()
-        if rest == "self_attention.linear_proj.weight":
-            return [(f"model.layers.{layer_idx}.self_attn.o_proj.weight", param)]
-        elif rest == "self_attention.linear_qkv.weight":
-
-            param = param.view(args.num_query_groups, -1, head_dim, args.hidden_size)
-            q_param, k_param, v_param = torch.split(param, split_size_or_sections=[value_num_per_group, 1, 1], dim=1)
-            q_param = q_param.reshape(-1, args.hidden_size)
-            k_param = k_param.reshape(-1, args.hidden_size)
-            v_param = v_param.reshape(-1, args.hidden_size)
-            return [
-                (f"model.layers.{layer_idx}.self_attn.q_proj.weight", q_param),
-                (f"model.layers.{layer_idx}.self_attn.k_proj.weight", k_param),
-                (f"model.layers.{layer_idx}.self_attn.v_proj.weight", v_param),
-            ]
-        elif rest == "self_attention.linear_qkv.bias":
-            param = param.view(args.num_query_groups, -1)
-            q_bias, k_bias, v_bias = torch.split(
-                param,
-                split_size_or_sections=[value_num_per_group * head_dim, head_dim, head_dim],
-                dim=1,
-            )
-            q_bias = q_bias.contiguous().flatten()
-            k_bias = k_bias.contiguous().flatten()
-            v_bias = v_bias.contiguous().flatten()
-            return [
-                (f"model.layers.{layer_idx}.self_attn.q_proj.bias", q_bias),
-                (f"model.layers.{layer_idx}.self_attn.k_proj.bias", k_bias),
-                (f"model.layers.{layer_idx}.self_attn.v_proj.bias", v_bias),
-            ]
-        elif rest == "mlp.linear_fc1.weight":
-            gate_weight, up_weight = param.chunk(2, dim=0)
-            return [
-                (f"model.layers.{layer_idx}.mlp.gate_proj.weight", gate_weight),
-                (f"model.layers.{layer_idx}.mlp.up_proj.weight", up_weight),
-            ]
-        elif rest == "mlp.linear_fc2.weight":
-            return [(f"model.layers.{layer_idx}.mlp.down_proj.weight", param)]
-        elif rest == "self_attention.linear_qkv.layer_norm_weight":
-            return [(f"model.layers.{layer_idx}.input_layernorm.weight", param)]
-        elif rest == "mlp.linear_fc1.layer_norm_weight":
-            return [(f"model.layers.{layer_idx}.post_attention_layernorm.weight", param)]
-
-        # qk norm
-        elif rest == "self_attention.q_layernorm.weight":
-            return [(f"model.layers.{layer_idx}.self_attn.q_norm.weight", param)]
-        elif rest == "self_attention.k_layernorm.weight":
-            return [(f"model.layers.{layer_idx}.self_attn.k_norm.weight", param)]
-
-    raise ValueError(f"Unknown parameter name: {name}")
-
-
-def convert_qwen3moe_to_hf(args, name, param):
-    if name == "module.module.embedding.word_embeddings.weight":
-        return [("model.embed_tokens.weight", param)]
-    if name == "module.module.output_layer.weight":
-        return [("lm_head.weight", param)]
-    if name == "module.module.decoder.final_layernorm.weight":
-        return [("model.norm.weight", param)]
-
-    try:
-        head_dim = args.kv_channels if args.kv_channels is not None else args.hidden_size // args.num_attention_heads
-    except:
-        head_dim = args.hidden_size // args.num_attention_heads
-    value_num_per_group = args.num_attention_heads // args.num_query_groups
-
-    decoder_layers_pattern = r"module\.module\.decoder\.layers\.(\d+)\.(.+)"
-    match = re.match(decoder_layers_pattern, name)
-    if match:
-        layer_idx, rest = match.groups()
-
-        # experts
-        expert_pattern = r"mlp.experts\.(.+)\.weight(\d+)"
-        match = re.match(expert_pattern, rest)
-        if match:
-            rest, expert_idx = match.groups()
-            if rest == "linear_fc1":
-                gate_weight, up_weight = param.chunk(2, dim=0)
-                outputs = [
-                    (f"model.layers.{layer_idx}.mlp.experts.{expert_idx}.gate_proj.weight", gate_weight),
-                    (f"model.layers.{layer_idx}.mlp.experts.{expert_idx}.up_proj.weight", up_weight),
-                ]
-                return outputs
-            elif rest == "linear_fc2":
-                outputs = [
-                    (f"model.layers.{layer_idx}.mlp.experts.{expert_idx}.down_proj.weight", param),
-                ]
-                if args.sglang_enable_ep_moe:
-                    outputs += [
-                        (
-                            f"model.layers.{layer_idx}.mlp.experts.{expert_idx}.down_proj.input_scale",
-                            torch.tensor(1.0, dtype=torch.float32, device=param.device),
-                        ),
-                        (
-                            f"model.layers.{layer_idx}.mlp.experts.{expert_idx}.down_proj.weight_scale",
-                            torch.tensor(1.0, dtype=torch.float32, device=param.device),
-                        ),
-                    ]
-                return outputs
-            else:
-                raise ValueError(f"Unknown expert parameter name: {name}")
-
-        # shared expert
-        shared_expert_pattern = r"mlp.shared_experts\.(.+)"
-        match = re.match(shared_expert_pattern, rest)
-        if match:
-            rest = match.groups()[0]
-            if rest == "linear_fc1.weight":
-                gate_weight, up_weight = param.chunk(2, dim=0)
-                return [
-                    (f"model.layers.{layer_idx}.mlp.shared_experts.gate_proj.weight", gate_weight),
-                    (f"model.layers.{layer_idx}.mlp.shared_experts.up_proj.weight", up_weight),
-                ]
-            elif rest == "linear_fc2.weight":
-                return [(f"model.layers.{layer_idx}.mlp.shared_experts.down_proj.weight", param)]
-            else:
-                raise ValueError(f"Unknown shared expert parameter name: {name}")
-
-        if rest == "self_attention.linear_proj.weight":
-            return [(f"model.layers.{layer_idx}.self_attn.o_proj.weight", param)]
-        elif rest == "self_attention.linear_qkv.weight":
-
-            param = param.view(args.num_query_groups, -1, head_dim, args.hidden_size)
-            q_param, k_param, v_param = torch.split(param, split_size_or_sections=[value_num_per_group, 1, 1], dim=1)
-            q_param = q_param.reshape(-1, args.hidden_size)
-            k_param = k_param.reshape(-1, args.hidden_size)
-            v_param = v_param.reshape(-1, args.hidden_size)
-            return [
-                (f"model.layers.{layer_idx}.self_attn.q_proj.weight", q_param),
-                (f"model.layers.{layer_idx}.self_attn.k_proj.weight", k_param),
-                (f"model.layers.{layer_idx}.self_attn.v_proj.weight", v_param),
-            ]
-        elif rest == "self_attention.linear_qkv.bias":
-            param = param.view(args.num_query_groups, -1)
-            q_bias, k_bias, v_bias = torch.split(
-                param,
-                split_size_or_sections=[value_num_per_group * head_dim, head_dim, head_dim],
-                dim=1,
-            )
-            q_bias = q_bias.contiguous().flatten()
-            k_bias = k_bias.contiguous().flatten()
-            v_bias = v_bias.contiguous().flatten()
-            return [
-                (f"model.layers.{layer_idx}.self_attn.q_proj.bias", q_bias),
-                (f"model.layers.{layer_idx}.self_attn.k_proj.bias", k_bias),
-                (f"model.layers.{layer_idx}.self_attn.v_proj.bias", v_bias),
-            ]
-        elif rest == "mlp.linear_fc1.weight":
-            gate_weight, up_weight = param.chunk(2, dim=0)
-            return [
-                (f"model.layers.{layer_idx}.mlp.gate_proj.weight", gate_weight),
-                (f"model.layers.{layer_idx}.mlp.up_proj.weight", up_weight),
-            ]
-        elif rest == "mlp.linear_fc2.weight":
-            return [(f"model.layers.{layer_idx}.mlp.down_proj.weight", param)]
-        elif rest == "self_attention.linear_qkv.layer_norm_weight":
-            return [(f"model.layers.{layer_idx}.input_layernorm.weight", param)]
-        elif rest == "mlp.linear_fc1.layer_norm_weight":
-            return [(f"model.layers.{layer_idx}.post_attention_layernorm.weight", param)]
-        elif rest == "pre_mlp_layernorm.weight":
-            return [(f"model.layers.{layer_idx}.post_attention_layernorm.weight", param)]
-        elif rest == "mlp.router.weight":
-            return [(f"model.layers.{layer_idx}.mlp.gate.weight", param)]
-        elif rest == "mlp.router.expert_bias":
-            return [(f"model.layers.{layer_idx}.mlp.gate.e_score_correction_bias", param)]
-
-        # qk norm
-        elif rest == "self_attention.q_layernorm.weight":
-            return [(f"model.layers.{layer_idx}.self_attn.q_norm.weight", param)]
-        elif rest == "self_attention.k_layernorm.weight":
-            return [(f"model.layers.{layer_idx}.self_attn.k_norm.weight", param)]
-
-    raise ValueError(f"Unknown parameter name: {name}")
-
-
-def convert_deepseekv3_to_hf(args, name, param):
-    if name == "module.module.embedding.word_embeddings.weight":
-        return [("model.embed_tokens.weight", param)]
-    if name == "module.module.output_layer.weight":
-        return [("lm_head.weight", param)]
-    if name == "module.module.decoder.final_layernorm.weight":
-        return [("model.norm.weight", param)]
-
-    try:
-        head_dim = args.kv_channels if args.kv_channels is not None else args.hidden_size // args.num_attention_heads
-    except:
-        head_dim = args.hidden_size // args.num_attention_heads
-    value_num_per_group = args.num_attention_heads // args.num_query_groups
-
-    decoder_layers_pattern = r"module\.module\.decoder\.layers\.(\d+)\.(.+)"
-    match = re.match(decoder_layers_pattern, name)
-    if match:
-        layer_idx, rest = match.groups()
-
-        # experts
-        expert_pattern = r"mlp.experts\.(.+)\.weight(\d+)"
-        match = re.match(expert_pattern, rest)
-        if match:
-            rest, expert_idx = match.groups()
-            if rest == "linear_fc1":
-                gate_weight, up_weight = param.chunk(2, dim=0)
-                outputs = [
-                    (f"model.layers.{layer_idx}.mlp.experts.{expert_idx}.gate_proj.weight", gate_weight),
-                    (f"model.layers.{layer_idx}.mlp.experts.{expert_idx}.up_proj.weight", up_weight),
-                ]
-                return outputs
-            elif rest == "linear_fc2":
-                outputs = [
-                    (f"model.layers.{layer_idx}.mlp.experts.{expert_idx}.down_proj.weight", param),
-                ]
-                if args.sglang_enable_ep_moe:
-                    outputs += [
-                        (
-                            f"model.layers.{layer_idx}.mlp.experts.{expert_idx}.down_proj.input_scale",
-                            torch.tensor(1.0, dtype=torch.float32, device=param.device),
-                        ),
-                        (
-                            f"model.layers.{layer_idx}.mlp.experts.{expert_idx}.down_proj.weight_scale",
-                            torch.tensor(1.0, dtype=torch.float32, device=param.device),
-                        ),
-                    ]
-                return outputs
-            else:
-                raise ValueError(f"Unknown expert parameter name: {name}")
-
-        # shared expert
-        shared_expert_pattern = r"mlp.shared_experts\.(.+)"
-        match = re.match(shared_expert_pattern, rest)
-        if match:
-            rest = match.groups()[0]
-            if rest == "linear_fc1.weight":
-                gate_weight, up_weight = param.chunk(2, dim=0)
-                return [
-                    (f"model.layers.{layer_idx}.mlp.shared_experts.gate_proj.weight", gate_weight),
-                    (f"model.layers.{layer_idx}.mlp.shared_experts.up_proj.weight", up_weight),
-                ]
-            elif rest == "linear_fc2.weight":
-                return [(f"model.layers.{layer_idx}.mlp.shared_experts.down_proj.weight", param)]
-            else:
-                raise ValueError(f"Unknown shared expert parameter name: {name}")
-
-        if rest == "self_attention.linear_proj.weight":
-            return [(f"model.layers.{layer_idx}.self_attn.o_proj.weight", param)]
-        elif rest == "self_attention.linear_q_proj.weight":
-            return [(f"model.layers.{layer_idx}.self_attn.q_proj.weight", param)]
-        elif rest == "self_attention.linear_qkv.bias":
-            param = param.view(args.num_query_groups, -1)
-            q_bias, k_bias, v_bias = torch.split(
-                param,
-                split_size_or_sections=[value_num_per_group * head_dim, head_dim, head_dim],
-                dim=1,
-            )
-            q_bias = q_bias.contiguous().flatten()
-            k_bias = k_bias.contiguous().flatten()
-            v_bias = v_bias.contiguous().flatten()
-            return [
-                (f"model.layers.{layer_idx}.self_attn.q_proj.bias", q_bias),
-                (f"model.layers.{layer_idx}.self_attn.k_proj.bias", k_bias),
-                (f"model.layers.{layer_idx}.self_attn.v_proj.bias", v_bias),
-            ]
-        elif rest == "mlp.linear_fc1.weight":
-            gate_weight, up_weight = param.chunk(2, dim=0)
-            return [
-                (f"model.layers.{layer_idx}.mlp.gate_proj.weight", gate_weight),
-                (f"model.layers.{layer_idx}.mlp.up_proj.weight", up_weight),
-            ]
-        elif rest == "mlp.linear_fc2.weight":
-            return [(f"model.layers.{layer_idx}.mlp.down_proj.weight", param)]
-        elif rest == "self_attention.linear_qkv.layer_norm_weight" or rest == "input_layernorm.weight":
-            return [(f"model.layers.{layer_idx}.input_layernorm.weight", param)]
-        elif rest == "mlp.linear_fc1.layer_norm_weight":
-            return [(f"model.layers.{layer_idx}.post_attention_layernorm.weight", param)]
-        elif rest == "self_attention.linear_kv_down_proj.weight":
-            return [(f"model.layers.{layer_idx}.self_attn.kv_a_proj_with_mqa.weight", param)]
-        elif rest == "self_attention.linear_kv_up_proj.layer_norm_weight":
-            return [(f"model.layers.{layer_idx}.self_attn.kv_a_layernorm.weight", param)]
-        elif rest == "self_attention.linear_kv_up_proj.weight":
-            return [(f"model.layers.{layer_idx}.self_attn.kv_b_proj.weight", param)]
-        elif rest == "pre_mlp_layernorm.weight":
-            return [(f"model.layers.{layer_idx}.post_attention_layernorm.weight", param)]
-        elif rest == "mlp.router.weight":
-            return [(f"model.layers.{layer_idx}.mlp.gate.weight", param)]
-        elif rest == "mlp.router.expert_bias":
-            return [(f"model.layers.{layer_idx}.mlp.gate.e_score_correction_bias", param)]
-
-    raise ValueError(f"Unknown parameter name: {name}")
-
-
-def convert_llama_to_hf(args, name, param):
-    if name == "module.module.embedding.word_embeddings.weight":
-        return [("model.embed_tokens.weight", param)]
-    if name == "module.module.output_layer.weight":
-        return [("lm_head.weight", param)]
-    if name == "module.module.decoder.final_layernorm.weight":
-        return [("model.norm.weight", param)]
-
-    try:
-        head_dim = args.kv_channels if args.kv_channels is not None else args.hidden_size // args.num_attention_heads
-    except:
-        head_dim = args.hidden_size // args.num_attention_heads
-    value_num_per_group = args.num_attention_heads // args.num_query_groups
-
-    decoder_layers_pattern = r"module\.module\.decoder\.layers\.(\d+)\.(.+)"
-    match = re.match(decoder_layers_pattern, name)
-    if match:
-        layer_idx, rest = match.groups()
-        if rest == "self_attention.linear_proj.weight":
-            return [(f"model.layers.{layer_idx}.self_attn.o_proj.weight", param)]
-        elif rest == "self_attention.linear_qkv.weight":
-            # Split QKV weight for Llama
-            param = param.view(args.num_query_groups, -1, head_dim, args.hidden_size)
-            q_param, k_param, v_param = torch.split(param, split_size_or_sections=[value_num_per_group, 1, 1], dim=1)
-            q_param = q_param.reshape(-1, args.hidden_size)
-            k_param = k_param.reshape(-1, args.hidden_size)
-            v_param = v_param.reshape(-1, args.hidden_size)
-            return [
-                (f"model.layers.{layer_idx}.self_attn.q_proj.weight", q_param),
-                (f"model.layers.{layer_idx}.self_attn.k_proj.weight", k_param),
-                (f"model.layers.{layer_idx}.self_attn.v_proj.weight", v_param),
-            ]
-        elif rest == "mlp.linear_fc1.weight":
-            # Split gate and up projections for SwiGLU
-            gate_weight, up_weight = param.chunk(2, dim=0)
-            return [
-                (f"model.layers.{layer_idx}.mlp.gate_proj.weight", gate_weight),
-                (f"model.layers.{layer_idx}.mlp.up_proj.weight", up_weight),
-            ]
-        elif rest == "mlp.linear_fc2.weight":
-            return [(f"model.layers.{layer_idx}.mlp.down_proj.weight", param)]
-        elif rest == "self_attention.linear_qkv.layer_norm_weight":
-            return [(f"model.layers.{layer_idx}.input_layernorm.weight", param)]
-        elif rest == "mlp.linear_fc1.layer_norm_weight":
-            return [(f"model.layers.{layer_idx}.post_attention_layernorm.weight", param)]
-        elif rest == "pre_mlp_layernorm.weight":
-            return [(f"model.layers.{layer_idx}.post_attention_layernorm.weight", param)]
-
-    raise ValueError(f"Unknown parameter name: {name}")
